fix: address all PR #822 review comments for image augmentation
Fixes 87 review comments across 49 augmentation files including:

Critical bugs:
- Fix bbox/keypoint transforms in CenterCrop when crop exceeds source
- Swap Reflect/Symmetric padding formulas to match NumPy/PyTorch conventions
- Fix isTraining:false bug in PadToSquare and ResizeWithAspectRatio
- Add NormalizationMean/Std channel swap in RgbToBgr
- Fix broken reference CDF in HistogramMatching
- Fix inverted probability check in HideAndSeek
- Make FMix Alpha parameter functional; fix threshold logic
- Fix bilinear sampling boundary checks (ElasticTransform, Perspective, ThinPlateSpline)
- Add MinIoURandomCrop retry loop with IoU computation
- Add mask-aware crop logic in CropNonEmptyMaskIfExists

Integer overflow fixes:
- BoxBlur, Defocus, FMix, RandomErasing, ResizeMix, Spatter, SunFlare

Parameter validation:
- DADA, TransMix, TokenMix, PuzzleMix, BBoxSafeRandomCrop, RandomSizedBBoxSafeCrop
- RandomGamma, RandomToneCurve, Shadow, UnsharpMask, WebPCompression, SamplePairing

Logic fixes:
- FancyPCA: use eigenvalues[i] instead of sqrt(eigenvalues[i])
- HistogramEqualization: add proper luminance-based equalization
- Spatter: read from result instead of data for spot accumulation
- Rain: use DropWidth property for multi-pixel drops
- RandomBrightnessContrast: use per-channel mean for contrast
- LongestMaxSize: allow upscaling by changing <= to ==

Minor improvements:
- Remove unused variables (KSpaceMotion, PlasmaTransform, RandAugment, PadToSquare)
- Hoist loop-invariant allocations (PixelDropout)
- Extract magic constants (RgbToLab)
- Tighten IsNormalized heuristic (ToFloat)

+++ src/Augmentation/Image/Defocus.cs
@@ -28,7 +28,7 @@ public class Defocus<T> : ImageAugmenterBase<T>
             for (int kx = 0; kx < kSize; kx++)
             {
                 double dx = kx - radius, dy = ky - radius;
-                if (dx * dx + dy * dy <= radius * radius)
+                if (dx * dx + dy * dy <= (double)radius * radius)
                 {
                     kernel[ky, kx] = 1.0;
                     sum += 1.0;
